updated player classes
