+++ src/main/java/de/roleplay/backend/entitys/playerfigure/concreteplayerfigure/talent/TalentFighter.java
@@ -1,0 +1,7 @@
+package de.roleplay.backend.entitys.playerfigure.concreteplayerfigure.talent;
+
+public enum TalentFighter {
+    Akrobatik,
+    Athletics,
+    Survival
+}
