+++ src/main/java/de/roleplay/backend/entitys/playerfigure/concreteplayerfigure/talent/TalentWizard.java
@@ -1,4 +1,4 @@
-package de.roleplay.backend.playerfigure.concreteplayerfigure.talent;
+package de.roleplay.backend.entitys.playerfigure.concreteplayerfigure.talent;
 
 public enum TalentWizard {
     Medicine,
