+++ src/main/java/de/roleplay/backend/entitys/playerfigure/monster/Monster.java
@@ -1,0 +1,10 @@
+package de.roleplay.backend.entitys.playerfigure.monster;
+
+import de.roleplay.backend.entitys.playerfigure.PlayerFigure;
+import de.roleplay.backend.entitys.playerfigure.race.Race;
+
+public class Monster extends PlayerFigure {
+    public Monster(String name, Race race) {
+        super(name, race);
+    }
+}
