+++ src/main/java/de/roleplay/backend/entitys/playerfigure/race/Elf.java
@@ -1,4 +1,4 @@
-package de.roleplay.backend.playerfigure.race;
+package de.roleplay.backend.entitys.playerfigure.race;
 
 public class Elf extends Race {
     public Elf() {
