+++ src/main/java/de/roleplay/backend/entitys/playerfigure/race/Human.java
@@ -1,4 +1,4 @@
-package de.roleplay.backend.playerfigure.race;
+package de.roleplay.backend.entitys.playerfigure.race;
 
 public class Human extends Race {
     public Human() {
