--- src/main/java/de/roleplay/backend/playerfigure/CharacterClass.java
@@ -1,4 +1,0 @@
-package de.roleplay.backend.playerfigure;
-
-public class CharacterClass {
-}
