--- src/main/java/de/roleplay/backend/playerfigure/concreteplayerfigure/Fighter.java
@@ -1,16 +1,0 @@
-package de.roleplay.backend.playerfigure.concreteplayerfigure;
-
-import de.roleplay.backend.playerfigure.PlayerFigure;
-import de.roleplay.backend.playerfigure.concreteplayerfigure.talent.TalentFighter;
-import de.roleplay.backend.playerfigure.race.Race;
-
-public class Fighter extends PlayerFigure {
-
-    private TalentFighter talent;
-
-    public Fighter(String name, Race race) {
-        super(name, race);
-        setHealthPoints(10);
-        setAmorClass(13);
-    }
-}
