--- src/main/java/de/roleplay/backend/playerfigure/concreteplayerfigure/talent/TalentFighter.java
@@ -1,7 +1,0 @@
-package de.roleplay.backend.playerfigure.concreteplayerfigure.talent;
-
-public enum TalentFighter {
-    Akrobatik,
-    Athletics,
-    Survival
-}
