--- src/main/java/de/roleplay/backend/playerfigure/monster/Monster.java
@@ -1,4 +1,0 @@
-package de.roleplay.backend.playerfigure.monster;
-
-public class Monster {
-}
